+++ src/Microsoft.Diagnostics.Runtime/src/Linux/StreamAddressSpace.cs
@@ -16,12 +16,12 @@ namespace Microsoft.Diagnostics.Runtime.Linux
             _stream = stream;
         }
 
-        public long Length => _stream.Length;
+        public ulong Length => (uint)_stream.Length;
         public string Name => _stream.GetFilename() ?? _stream.GetType().Name;
 
-        public int Read(long position, Span<byte> buffer)
+        public int Read(ulong position, Span<byte> buffer)
         {
-            _stream.Seek(position, SeekOrigin.Begin);
+            _stream.Seek((long)position, SeekOrigin.Begin);
             return _stream.Read(buffer);
         }
     }
